Add JWT auth

+++ src/controllers/review.controller.ts
@@ -7,35 +7,37 @@ import { User } from '../models/user.model';
 import { setRestaurantAverageScoreAndPrice } from '../utils';
 
 export const createReview = async (req: Request<any, any, CreateReviewInputValidation['body']>, res: Response) => {
-  const { user: userInput, restaurant: restaurantInput, content, score, price } = req.body;
+  const { id: userId } = req.scopedInfo!;
+  const { restaurant: restaurantId, content, score, price } = req.body;
 
-  const user = await User.findById(userInput);
-  const restaurant = await Restaurant.findById(restaurantInput);
+  const restaurant = await Restaurant.findById(restaurantId);
 
-  if (any(isNil, [restaurant, user])) {
-    return res
-      .status(404)
-      .send({ message: `User with id "${userInput}" or Restaurant with id "${restaurantInput}" not found.` });
+  if (isNil(restaurant)) {
+    return res.status(404).send({ message: `Restaurant with id "${restaurantId}" not found.` });
   }
 
-  const review = await Review.findOne({ user: userInput, restaurant: restaurantInput });
+  const review = await Review.findOne({ user: userId, restaurant: restaurantId });
 
   if (!isNil(review)) {
     return res.status(422).send({ message: `Can't review a restaurant more than once. Update your review instead.` });
   }
 
-  const data = await Review.create({ content, score, price, user: userInput, restaurant: restaurantInput });
+  const data = await Review.create({ content, score, price, user: userId, restaurant: restaurantId });
 
-  await User.findByIdAndUpdate(userInput, { $push: { reviews: data.id } });
-  await Restaurant.findByIdAndUpdate(restaurantInput, { $push: { reviews: data.id } });
+  await User.findByIdAndUpdate(userId, { $push: { reviews: data.id } });
+  await Restaurant.findByIdAndUpdate(restaurantId, { $push: { reviews: data.id } });
 
-  await setRestaurantAverageScoreAndPrice(restaurantInput);
+  await setRestaurantAverageScoreAndPrice(restaurantId);
 
   return res.send({ data });
 };
 
 export const getAllReviews = async (_: Request, res: Response) => {
-  const data = await Review.find().sort('-updatedAt').populate(['user', 'restaurant']).exec();
+  const data = await Review.find()
+    .sort('-updatedAt')
+    .populate({ path: 'user', select: '-password' })
+    .populate('restaurant')
+    .exec();
 
   return res.send({ data });
 };
@@ -43,7 +45,7 @@ export const getAllReviews = async (_: Request, res: Response) => {
 export const getReview = async (req: Request, res: Response) => {
   const { id } = req.params;
 
-  const data = await Review.findById(id).populate(['user', 'restaurant']);
+  const data = await Review.findById(id).populate({ path: 'user', select: '-password' }).populate('restaurant');
 
   if (isNil(data)) {
     return res.status(404).send({ message: `Review with id "${id}" not found.` });
@@ -57,6 +59,7 @@ export const updateReview = async (
   res: Response
 ) => {
   const { id } = req.params;
+  const { id: userId } = req.scopedInfo!;
   const { content, score, price } = req.body;
 
   const review = await Review.findById(id);
@@ -65,7 +68,15 @@ export const updateReview = async (
     return res.status(404).send({ message: `Review with id "${id}" not found.` });
   }
 
-  await Review.findByIdAndUpdate(id, { content, score, price: price ?? review.price });
+  if (!review.user.equals(userId)) {
+    return res.status(422).send({ message: `Can't update review that you don't own.` });
+  }
+
+  await Review.findByIdAndUpdate(id, {
+    content: content ?? review.content,
+    score: score ?? review.score,
+    price: price ?? review.price
+  });
   await setRestaurantAverageScoreAndPrice(review.restaurant);
 
   const data = await Review.findById(id);
@@ -75,11 +86,16 @@ export const updateReview = async (
 
 export const deleteReview = async (req: Request, res: Response) => {
   const { id } = req.params;
+  const { id: userId } = req.scopedInfo!;
 
   const review = await Review.findById(id);
 
   if (isNil(review)) {
     return res.status(404).send({ message: `Review with id "${id}" not found.` });
+  }
+
+  if (!review.user.equals(userId)) {
+    return res.status(422).send({ message: `Can't delete a review that you don't own.` });
   }
 
   await Review.findByIdAndDelete(id);
